Add comments to functions and move interfaces to libs
