+++ libs/vscode-extension/src/lib/models/templates/IBindingTemplate.ts
@@ -8,7 +8,6 @@ export type BindingSettingValue = string | boolean | number | undefined;
  * Describes a template used for creating a binding
  */
 export interface IBindingTemplate {
-  // tslint:disable-next-line: no-reserved-keywords
   type: string;
   direction: string;
   displayName: string;
